requirements.txt
# local package
-e .
pandas
numpy
# external requirements
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1
